map_backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
httpx
python-dotenv
pydantic
aiofiles
